Makefile.am: Do not build dbl-64 variants.
2006-03-30  Richard Guenther  <rguenther@suse.de>

	* i386/Makefile.am: Do not build dbl-64 variants.
	* i386/Makefile.in: Regenerate.
	* dbl-64: Remove.

	* config/i386/i386.c: Remove builtins for SSE2 ABI intrinsic
	variants with double arguments.

	* gcc.target/i386/sselibm-1.c: Adjust for libgcc-math partial
	removal.
	* gcc.target/i386/sselibm-2.c: Likewise.
	* gcc.target/i386/sselibm-3.c: Likewise.
	* gcc.target/i386/sselibm-4.c: Likewise.
	* gcc.target/i386/sselibm-5.c: Likewise.

From-SVN: r112530

+++ gcc/testsuite/gcc.target/i386/sselibm-1.c
@@ -2,11 +2,11 @@
 /* { dg-options "-msse2 -mfpmath=sse" } */
 /* { dg-require-effective-target ilp32 } */
 
-double sin(double);
+float sinf(float);
 
-double foo(double x)
+float foo(float x)
 {
-  return sin(x);
+  return sinf(x);
 }
 
-/* { dg-final { scan-assembler-not "__libm_sse2_sin" } } */
+/* { dg-final { scan-assembler-not "__libm_sse2_sinf" } } */
